Remove Read Handler Cache
Finding an entry in the read handler cache is very expensive if the map
of custom handlers is big. For single reads per reader the overhead of
using the cache is substantial. I can show, that it is faster to just
merge a new map of default handlers with the custom handlers every time.

To minimize the size of default handlers and avoid the cost of checking
for overridden ground types, I use a switch statement for ground types
instead of using the normal map lookup.

The only outside facing change is the lack of error reporting in case a
ground type is overridden.