Make ReaderFactory an Instance holding the Handlers
ReaderFactory was a static class using a cache of handlers. However if
one has many custom handlers, the cache lookup is very expensive. I
added public methods to create a ReaderFactory in advance of creating a
reader which allows for handler map sharing.

The original methods create a ReaderFactory every time and loose the
caching functionality.